pyproject.toml
[tool.poetry]
name = "experiments"
version = "0.1.0"
description = ""
authors = ["Clay McCullough <cmckilla44@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.8"
python-dotenv = "^1.0.1"
jq = "^1.6.0"
langchain-mistralai = "^0.0.4"
pypdf = "^4.0.2"
#pysqlite3-binary = "^0.5.2.post3"
chromadb = "^0.4.22"
numexpr = "^2.9.0"
openai = "^0.28.0"
streamlit = "^1.31.1"
langchainhub = "^0.1.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
